Created dashboards for each team, added UserPage to display the dashboard of favorite team,
Every team has their own dashboard now. When a user creates an account, they are asked to choose their favorite team, When logged in, the dashboard of the team they chose is displayed, with the option to go to the original dashboard page, which contains all of the team's statistics. Each dashboard uses the color scheme of the franchise and contains both current and predictive data.