Fixed import issue, removed sensitive from previous commits, added docker instructions to readme.
The passwords themselves were expired, but we utilized git filter-repo anyways to remove any previous sensitive data from the repository, most of which belonging to .env files.
Therefore, in order to run the docker container, you must specify the database credentials in the .env file to ensure proper authentication when altering the current_data_table.